+++ src/gui/alphasetup.cpp
@@ -103,7 +103,7 @@ int CAlphaSetup::exec(CMenuTarget* parent, const std::string &)
 	int selected = 0;
 //	int max = 1;
 
-	unsigned long long timeoutEnd = CRCInput::calcTimeoutEnd(g_settings.timing[SNeutrinoSettings::TIMING_MENU] == 0 ? 0xFFFF : g_settings.timing[SNeutrinoSettings
+	uint64_t timeoutEnd = CRCInput::calcTimeoutEnd(g_settings.timing[SNeutrinoSettings::TIMING_MENU] == 0 ? 0xFFFF : g_settings.timing[SNeutrinoSettings
 ::TIMING_MENU]);
 
 	bool loop=true;
